+++ grails-app/conf/spring/resources.groovy
@@ -1,6 +1,2 @@
-import zelifkudos.ChatWebSocketController
-
-// Place your Spring DSL code here
 beans = {
-    chatWebSocketController(ChatWebSocketController)
 }