+++ grails-app/domain/kudos/Kudos.groovy
@@ -1,0 +1,15 @@
+package kudos
+
+class Kudos {
+
+    User sender
+    User receiver
+    String message
+    Date dateCreated
+
+    static constraints = {
+        sender nullable: false
+        receiver nullable: false
+        message nullable: true, maxSize: 200
+    }
+}