+++ src/main/groovy/zelifkudos/KudosLimitException.groovy
@@ -1,7 +1,0 @@
-package zelifkudos
-
-class KudosLimitException extends Exception {
-    KudosLimitException(String message) {
-        super(message)
-    }
-}